+++ ui/sxgo_doc.hpp
@@ -17,7 +17,7 @@ public:
 	virtual wxInputStream& LoadObject(wxInputStream& stream);
 	listing_info GetListing() const;
 	unsigned short SingleStep();
-	unsigned short Run( unsigned long count);
+	unsigned long Run( unsigned long count);
 	void SetBreakpoint( unsigned short address, bool do_set = true);
 	sx_simulator::state GetState() const;
